Refactoring and code cleanup for data paging feature
