+++ app/redux/constants/actionTypes.js
@@ -240,8 +240,3 @@ export const GENERATE_PDF_SUCCESS = 'GENERATE_PDF_SUCCESS';
 export const GENERATE_PDF_FAILURE = 'GENERATE_PDF_FAILURE';
 
 export const REMOVE_GENERATED_PDFS = 'REMOVE_GENERATED_PDFS';
-
-// data processing
-export const PROCESS_PATIENT_DATA_REQUEST = 'PROCESS_PATIENT_DATA_REQUEST';
-export const PROCESS_PATIENT_DATA_SUCCESS = 'PROCESS_PATIENT_DATA_SUCCESS';
-export const PROCESS_PATIENT_DATA_FAILURE = 'PROCESS_PATIENT_DATA_FAILURE';
