+++ app/redux/selectors/index.js
@@ -1,24 +1,5 @@
-import { createSelector } from 'reselect';
-import _ from 'lodash';
-import utils from '../../core/utils'
+import { getfetchedPatientDataRange } from './patientData';
 
-const getPatientData = (state, props) => _.get(
-  state,
-  `blip.patientDataMap[${_.get(props, 'routeParams.id')}]`,
-  []
-);
-
-const getPatientFetchedUntil = (state, props) => _.get(
-  state,
-  `blip.patientDataMap[${_.get(props, 'routeParams.id')}_fetchedUntil]`
-);
-
-export const getfetchedPatientDataRange = createSelector(
-  [ getPatientData, getPatientFetchedUntil ],
-  (data, fetchedUntil) =>  _.assign({}, utils.getDeviceDataRange(data), {
-    fetchedUntil,
-    count: data ? data.length : 0,
-  }),
-);
-
-export default getfetchedPatientDataRange;
+export {
+  getfetchedPatientDataRange,
+}
