+++ package.json
@@ -76,7 +76,6 @@
     "tideline": "0.6.14",
     "tidepool-platform-client": "0.34.3",
     "uglify-es": "3.0.13",
-    "universe": "0.8.1",
     "url-loader": "0.5.7",
     "webpack": "1.13.2",
     "worker-loader": "0.8.0"
